+++ keyboards/designedbygg/ironclad/icl01/keymaps/via/ironclad_via.json
@@ -1,5 +1,5 @@
 {
-  "name": "DesignedbyGG IronClad",
+  "name": "DesignedbyGG IronClad ICL01",
   "vendorId": "0x320F",
   "productId": "0x5041",
   "keycodes": ["qmk_lighting"],
