+++ keyboards/designedbygg/ironclad/icl03/icl03.c
@@ -14,12 +14,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#pragma once
-#define RGB_MATRIX_DEFAULT_MODE RGB_MATRIX_SOLID_REACTIVE_SIMPLE;
-#define WEAR_LEVELING_SN32_EMULATION_PAGE_COUNT 32
-// Enable RGB Matrix
-#define VIA_QMK_RGBLIGHT_ENABLE
-#define DYNAMIC_KEYMAP_LAYER_COUNT 3
-#define CAPS_LOCK_INDEX 61
-#define NUM_LOCK_INDEX 37
-//#define EECONFIG_KB_DATA_SIZE 16
+#include "quantum.h"
+#ifdef LED_WIN_LOCK_PIN
+/* Handle the Win Lock LED */
+void keyboard_pre_init_kb(void) {
+	setPinOutput(LED_WIN_LOCK_PIN);
+	writePin(LED_WIN_LOCK_PIN, !LED_PIN_ON_STATE);
+}
+
+void housekeeping_task_kb(void) {
+	writePin(LED_WIN_LOCK_PIN,!keymap_config.no_gui);
+}
+#endif
