+++ keyboards/designedbygg/ironclad/icl03/keymaps/via/rules.mk
@@ -1,7 +1,7 @@
 VIA_ENABLE = yes
-#MOUSEKEY_ENABLE = yes
 LED_WIN_LOCK_PIN = yes
 
+# Reduce RAM usage
 ifeq ($(strip $(RGB_MATRIX_ENABLE)), yes)
 LTO_ENABLE = yes
 endif
