+++ keyboards/designedbygg/ironclad/icl03/readme.md
@@ -1,6 +1,6 @@
-# GG Berserker
+# GG Ironclad
 
-Berserker
+Ironclad ICL03
 
 Keyboard Maintainer: [Dimitris Mantzouranis](https://github.com/dexter93)
 Brought to you by: [SonixQMK](https://github.com/SonixQMK/qmk_firmware)
@@ -8,6 +8,6 @@ Hardware Supported: SN32F248BF
 
 Make example for this keyboard (after setting up your build environment):
 
-    qmk compile -kb designedbygg/berserker -km default
+    qmk compile -kb designedbygg/ironclad/icl03 -km default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
